Bump to the latest contact form.

+++ wp-content/plugins/contact-form-7/includes/functions.php
@@ -7,7 +7,8 @@ function wpcf7_plugin_path( $path = '' ) {
 function wpcf7_plugin_url( $path = '' ) {
 	$url = plugins_url( $path, WPCF7_PLUGIN );
 
-	if ( is_ssl() && 'http:' == substr( $url, 0, 5 ) ) {
+	if ( is_ssl()
+	and 'http:' == substr( $url, 0, 5 ) ) {
 		$url = 'https:' . substr( $url, 5 );
 	}
 
@@ -19,7 +20,8 @@ function wpcf7_upload_dir( $type = false ) {
 
 	$uploads = apply_filters( 'wpcf7_upload_dir', array(
 		'dir' => $uploads['basedir'],
-		'url' => $uploads['baseurl'] ) );
+		'url' => $uploads['baseurl'],
+	) );
 
 	if ( 'dir' == $type ) {
 		return $uploads['dir'];
@@ -30,16 +32,12 @@ function wpcf7_upload_dir( $type = false ) {
 	return $uploads;
 }
 
-function wpcf7_verify_nonce( $nonce, $action = -1 ) {
-	if ( substr( wp_hash( $action, 'nonce' ), -12, 10 ) == $nonce ) {
-		return true;
-	}
-
-	return false;
+function wpcf7_verify_nonce( $nonce, $action = 'wp_rest' ) {
+	return wp_verify_nonce( $nonce, $action );
 }
 
-function wpcf7_create_nonce( $action = -1 ) {
-	return substr( wp_hash( $action, 'nonce' ), -12, 10 );
+function wpcf7_create_nonce( $action = 'wp_rest' ) {
+	return wp_create_nonce( $action );
 }
 
 function wpcf7_blacklist_check( $target ) {
@@ -54,7 +52,8 @@ function wpcf7_blacklist_check( $target ) {
 	foreach ( (array) $words as $word ) {
 		$word = trim( $word );
 
-		if ( empty( $word ) || 256 < strlen( $word ) ) {
+		if ( empty( $word )
+		or 256 < strlen( $word ) ) {
 			continue;
 		}
 
@@ -109,6 +108,10 @@ function wpcf7_use_really_simple_captcha() {
 function wpcf7_validate_configuration() {
 	return apply_filters( 'wpcf7_validate_configuration',
 		WPCF7_VALIDATE_CONFIGURATION );
+}
+
+function wpcf7_autop_or_not() {
+	return (bool) apply_filters( 'wpcf7_autop_or_not', WPCF7_AUTOP );
 }
 
 function wpcf7_load_js() {
@@ -191,7 +194,8 @@ function wpcf7_register_post_types() {
 function wpcf7_version( $args = '' ) {
 	$defaults = array(
 		'limit' => -1,
-		'only_major' => false );
+		'only_major' => false,
+	);
 
 	$args = wp_parse_args( $args, $defaults );
 
@@ -251,26 +255,33 @@ function wpcf7_enctype_value( $enctype ) {
 
 function wpcf7_rmdir_p( $dir ) {
 	if ( is_file( $dir ) ) {
-		if ( ! $result = @unlink( $dir ) ) {
-			$stat = @stat( $dir );
-			$perms = $stat['mode'];
-			@chmod( $dir, $perms | 0200 ); // add write for owner
+		$file = $dir;
 
-			if ( ! $result = @unlink( $dir ) ) {
-				@chmod( $dir, $perms );
-			}
+		if ( @unlink( $file ) ) {
+			return true;
 		}
 
-		return $result;
+		$stat = stat( $file );
+
+		if ( @chmod( $file, $stat['mode'] | 0200 ) ) { // add write for owner
+			if ( @unlink( $file ) ) {
+				return true;
+			}
+
+			@chmod( $file, $stat['mode'] );
+		}
+
+		return false;
 	}
 
 	if ( ! is_dir( $dir ) ) {
 		return false;
 	}
 
-	if ( $handle = @opendir( $dir ) ) {
+	if ( $handle = opendir( $dir ) ) {
 		while ( false !== ( $file = readdir( $handle ) ) ) {
-			if ( $file == "." || $file == ".." ) {
+			if ( $file == "."
+			or $file == ".." ) {
 				continue;
 			}
 
@@ -280,7 +291,12 @@ function wpcf7_rmdir_p( $dir ) {
 		closedir( $handle );
 	}
 
-	return @rmdir( $dir );
+	if ( false !== ( $files = scandir( $dir ) )
+	and ! array_diff( $files, array( '.', '..' ) ) ) {
+		return rmdir( $dir );
+	}
+
+	return false;
 }
 
 /* From _http_build_query in wp-includes/functions.php */
@@ -301,7 +317,7 @@ function wpcf7_build_query( $args, $key = '' ) {
 			$v = '0';
 		}
 
-		if ( is_array( $v ) || is_object( $v ) ) {
+		if ( is_array( $v ) or is_object( $v ) ) {
 			array_push( $ret, wpcf7_build_query( $v, $k ) );
 		} else {
 			array_push( $ret, $k . '=' . urlencode( $v ) );
@@ -330,6 +346,7 @@ function wpcf7_count_code_units( $string ) {
 	}
 
 	$string = (string) $string;
+	$string = str_replace( "\r\n", "\n", $string );
 
 	$encoding = mb_detect_encoding( $string, mb_detect_order(), true );
 
@@ -352,11 +369,65 @@ function wpcf7_is_localhost() {
 function wpcf7_deprecated_function( $function, $version, $replacement ) {
 	$trigger_error = apply_filters( 'deprecated_function_trigger_error', true );
 
-	if ( WP_DEBUG && $trigger_error ) {
+	if ( WP_DEBUG and $trigger_error ) {
 		if ( function_exists( '__' ) ) {
 			trigger_error( sprintf( __( '%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.', 'contact-form-7' ), $function, $version, $replacement ) );
 		} else {
 			trigger_error( sprintf( '%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.', $function, $version, $replacement ) );
 		}
 	}
+}
+
+function wpcf7_log_remote_request( $url, $request, $response ) {
+	$log = sprintf(
+		/* translators: 1: response code, 2: response message, 3: URL */
+		__( 'HTTP Response: %1$s %2$s from %3$s', 'contact-form-7' ),
+		(int) wp_remote_retrieve_response_code( $response ),
+		wp_remote_retrieve_response_message( $response ),
+		$url
+	);
+
+	$log = apply_filters( 'wpcf7_log_remote_request',
+		$log, $url, $request, $response
+	);
+
+	if ( $log ) {
+		trigger_error( $log );
+	}
+}
+
+function wpcf7_anonymize_ip_addr( $ip_addr ) {
+	if ( ! function_exists( 'inet_ntop' )
+	or ! function_exists( 'inet_pton' ) ) {
+		return $ip_addr;
+	}
+
+	$packed = inet_pton( $ip_addr );
+
+	if ( false === $packed ) {
+		return $ip_addr;
+	}
+
+	if ( 4 == strlen( $packed ) ) { // IPv4
+		$mask = '255.255.255.0';
+	} elseif ( 16 == strlen( $packed ) ) { // IPv6
+		$mask = 'ffff:ffff:ffff:0000:0000:0000:0000:0000';
+	} else {
+		return $ip_addr;
+	}
+
+	return inet_ntop( $packed & inet_pton( $mask ) );
+}
+
+function wpcf7_is_file_path_in_content_dir( $path ) {
+	if ( 0 === strpos( realpath( $path ), realpath( WP_CONTENT_DIR ) ) ) {
+		return true;
+	}
+
+	if ( defined( 'UPLOADS' )
+	and 0 === strpos( realpath( $path ), realpath( ABSPATH . UPLOADS ) ) ) {
+		return true;
+	}
+
+	return false;
 }
